Fix custom reporter test on PHP 5.6
